app/requirements_app.txt
beautifulsoup4
certifi
click
colorama
discord
distro
elasticsearch
fastapi
flask
flask-cors
h11
httpcore
httpx
idna
jinja2
jiter
JPype1
konlpy
MarkupSafe
numpy
openai
pandas
passlib[bcrypt]
plotly
pydantic
pyjwt
pymongo
python-dotenv
python-jose[cryptography]
python-multipart
python-telegram-bot
requests
sniffio
starlette
torch
tqdm
transformers
typing-inspection
typing_extensions
uvicorn[standard]
yfinance